+++ protected/humhub/modules/file/components/FileManager.php
@@ -8,6 +8,7 @@
 
 namespace humhub\modules\file\components;
 
+use humhub\modules\search\libs\SearchHelper;
 use Yii;
 use yii\base\Component;
 use humhub\modules\file\models\File;
@@ -30,7 +31,7 @@ class FileManager extends Component
     /**
      * Attach files to record.
      * This is required when uploaded before the related content is saved.
-     * 
+     *
      * @param string|array|File $files of File records or comma separeted list of file guids or single File record
      * @param boolean $steal steal when already assigned to other record
      */
@@ -66,13 +67,12 @@ class FileManager extends Component
             ]);
         }
 
-        // Save record to force search index update
-        $this->record->save();
+        SearchHelper::queueUpdate($this->record);
     }
 
     /**
      * File find query
-     * 
+     *
      * @return \yii\db\ActiveQuery file find query
      */
     public function find()
@@ -82,8 +82,8 @@ class FileManager extends Component
 
     /**
      * Returns a list of files assigned to the record
-     * 
-     * @return File[] array of files assigned to the record 
+     *
+     * @return File[] array of files assigned to the record
      */
     public function findAll()
     {