+++ protected/humhub/modules/ui/config.php
@@ -1,7 +1,13 @@
 <?php
 
+use humhub\components\console\Application;
+
+/** @noinspection MissedFieldInspection */
 return [
     'id' => 'ui',
     'class' => \humhub\modules\ui\Module::class,
     'isCoreModule' => true,
+    'events' => [
+        [Application::class, Application::EVENT_ON_INIT, ['humhub\modules\ui\Events', 'onConsoleApplicationInit']],
+    ]
 ];